Add files via upload
Final Commit, changes include .mp3 file for audio player feature (multimedia feature), and three .jpeg files for "Meet the Team" feature. Two other changes made to the website include the following: social media icons in which clients can interact with at the bottom of the page, and user login feature (text inputs for username and password, and button to enter client input). It is worth noting that these features are not exactly responsive the way a website would typically operate.